Split out immunize_method into its own file
Note: immunize_method will allow caps lock to be used/toggled, but
"immunize" its effects. In other words, when caps lock is on hitting a
letter on your keyboard generates a lowercase keypress unless shift is
held, as if caps lock wasn't on.

This method works well in desktop/office type applications, but does not
play well with games for many reasons.